Move backend peripherals tests
